requirements.txt
# OmniMe - macOS 输入追踪系统
# Python 3.10+

# macOS 原生 API 绑定
pyobjc-core>=10.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
pyobjc-framework-ApplicationServices>=10.0

# 数据库
# SQLite 是 Python 内置的，无需额外安装

# UI 相关
rumps>=0.4.0  # macOS Menu Bar 应用框架

# Web 后台
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# 数据处理
python-dateutil>=2.8.2

# AI 总结（可选）
openai>=1.0.0

# 开发工具
rich>=13.0.0  # 美化终端输出
